* Load vocab mappings in separate class (instantiate once). * Enable loading filter mappings from AppParams (beside vocab mappings) into Disambiguate. * Move disambiguation stuff from QueryTreeBuiler to Disambiguate.
